category update

+++ src/main/java/com/quynhtd/source_code_final/service/CategoryService.java
@@ -1,6 +1,7 @@
 package com.quynhtd.source_code_final.service;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -10,22 +11,34 @@ import com.quynhtd.source_code_final.repository.CategoryRepository;
 
 @Service
 public class CategoryService {
-    @Autowired
-    private CategoryRepository categoryRepository;
+	  @Autowired
+	    private CategoryRepository categoryRepository;
 
-    public List<Category> getAllCategorys() {
-        return categoryRepository.findAll();
-    }
+	    public List<Category> getAllCategorys() {
+	        return categoryRepository.findAll();
+	    }
 
-    public Category getCategoryById(Long id) {
-        return categoryRepository.findById(id).orElse(null);
-    }
+	    public Optional<Category> getCategoryById(Long id) {
+	        return categoryRepository.findById(id);
+	    }
 
-    public Category saveCategory(Category category) {
-        return categoryRepository.save(category);
-    }
+	    public Category createCategory(Category category) {
+	        return categoryRepository.save(category);
+	    }
 
-    public void deleteCategory(Long id) {
-        categoryRepository.deleteById(id);
-    }
-}
+	    public Category updateCategory(Long id, Category category) {
+	        if (categoryRepository.existsById(id)) {
+	            category.setId(id);
+	            return categoryRepository.save(category);
+	        }
+	        return null;
+	    }
+
+	    public boolean deleteCategory(Long id) {
+	        if (categoryRepository.existsById(id)) {
+	            categoryRepository.deleteById(id);
+	            return true;
+	        }
+	        return false;
+	    }
+	}
